requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
icalendar>=5.0.0
pytz>=2024.1
python-dateutil>=2.8.0
PyJWT>=2.8.0
cryptography>=42.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
python-multipart>=0.0.9
apscheduler>=3.10.0
slowapi>=0.1.9
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=2.0.0
